+++ src/components/BadgerBadge/stories.js
@@ -14,9 +14,18 @@ const defaultOpReturn = [
 	'Try out Badger at https://badger.bitcoin.com',
 ];
 
+const coinTypeOptions = ['BCH', 'SLP'];
+
+// [ NAKAMOTO, DOGECASH, BROC ]
+const tokenIdOptions = [
+	'df808a41672a0a0ae6475b44f272a107bc9961b90f29dc918d71301f24fe92fb',
+	'3916a24a051f8b3833a7fd128be51dd93015555ed9142d6106ec03267f5cdc4c',
+	'259908ae44f46ef585edef4bcc1e50dc06e4c391ac4be929fae27235b8158cf1',
+];
+
 storiesOf('BadgerBadge', module)
 	.add(
-		'all knobs',
+		'most knobs',
 		() => (
 			<BadgerBadge
 				price={number('Price', 0.001)}
@@ -25,14 +34,14 @@ storiesOf('BadgerBadge', module)
 					'To Address',
 					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
 				)}
-				opReturn={array('OP_RETURN', defaultOpReturn)}
+				opReturn={array('OP_RETURN', [])}
 				tag={text('Button Text', 'Badger Pay')}
 				text={text('Top Text', 'Payment Total')}
 				isRepeatable={boolean('Repeatable payment', false)}
 				repeatTimeout={number('Repeat Timeout (ms)', 4000)}
 				watchAddress={boolean('Watch Address All', true)}
 				showBrand={boolean('Toggle Brand', false)}
-				showSatoshis={boolean('Toggle Satoshis', true)}
+				showAmount={boolean('Toggle Amount', true)}
 				showQR={boolean('Toggle QR', true)}
 				showBorder={boolean('Toggle Border', false)}
 				successFn={() => console.log('success')}
@@ -41,7 +50,7 @@ storiesOf('BadgerBadge', module)
 		),
 		{
 			notes:
-				'Badger Badges are perfect for showing the price and Satoshis in a simple clean all in one component.  Default has knobs to experiment with all settings',
+				'Badger Badges are perfect for showing the price and amount in a simple clean all in one component.  Default has knobs to experiment with all settings',
 		}
 	)
 	.add(
@@ -54,21 +63,38 @@ storiesOf('BadgerBadge', module)
 					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
 				)}
 				text={''}
-				showSatoshis={false}
+				showAmount={false}
 				showQR={boolean('Toggle QR', true)}
 			/>
 		),
 		{
-			notes:
-				'Minimal look of Badge',
+			notes: 'Minimal look of Badge',
+		}
+	)
+	.add(
+		'price in fiat',
+		() => (
+			<BadgerBadge
+				price={number('Price', 0.001)}
+				currency={select('Currency', currencyOptions, 'USD')}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+				successFn={() => console.log('success')}
+				failFn={() => console.log('fail')}
+			/>
+		),
+		{
+			notes: 'Pay in any currency, and automagically convert the amount to BCH',
 		}
 	)
 	.add(
 		'price in BCH',
 		() => (
 			<BadgerBadge
-				ticker="BCH"
-				amount={number('BCH Amount', 0.0001)}
+				coinType="BCH"
+				amount={number('Amount', 0.0001)}
 				to={text(
 					'To Address',
 					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
@@ -76,8 +102,29 @@ storiesOf('BadgerBadge', module)
 			/>
 		),
 		{
-			notes:
-				'Badger Badges are perfect for showing the price and Satoshis in a simple clean all in one component.  Default has knobs to experiment with all settings',
+			notes: 'Price in BCH absolute value',
+		}
+	)
+	.add(
+		'price in SLP tokens',
+		() => (
+			<BadgerBadge
+				to={text(
+					'To Address',
+					'simpleledger:qq6qcjt6xlkeqzdwkhdvfyl2q2d2wafkgg8phzcqez'
+				)}
+				coinType="SLP"
+				tokenId={
+					text('Token ID', '') ||
+					select('Token ID Select', tokenIdOptions, tokenIdOptions[0])
+				}
+				amount={number('Amount', 5)}
+				tag="Send Tokens"
+				text="Send SLP Tokens"
+			/>
+		),
+		{
+			notes: 'Enter the token ID and send whichever SLP tokens you want!',
 		}
 	)
 	.add(
@@ -101,24 +148,6 @@ storiesOf('BadgerBadge', module)
 		}
 	)
 	.add(
-		'currency variety',
-		() => (
-			<BadgerBadge
-				price={number('Price', 0.001)}
-				currency={select('Currency', currencyOptions, 'USD')}
-				to={text(
-					'To Address',
-					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
-				)}
-				successFn={() => console.log('success')}
-				failFn={() => console.log('fail')}
-			/>
-		),
-		{
-			notes: 'Pay in any currency, and automagically convert the amount to BCH',
-		}
-	)
-	.add(
 		'toggle QR code',
 		() => (
 			<BadgerBadge
@@ -134,11 +163,11 @@ storiesOf('BadgerBadge', module)
 		),
 		{
 			notes:
-				'Badger Badges are perfect for showing the price and Satoshis in a simple clean all in one component.  Default has knobs to experiment with all settings',
+				'Optional QR code in addition to Button.  Only shows if transaction fully compatible in a URI',
 		}
 	)
 	.add(
-		'toggle satoshis',
+		'toggle coin amount',
 		() => (
 			<BadgerBadge
 				price={0.001}
@@ -146,13 +175,13 @@ storiesOf('BadgerBadge', module)
 					'To Address',
 					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
 				)}
-				showSatoshis={boolean('Show Satoshis', false)}
+				showAmount={boolean('Show Satoshis', false)}
 				successFn={() => console.log('success')}
 				failFn={() => console.log('fail')}
 			/>
 		),
 		{
-			notes: 'Choose to show the Satoshi amount alongside the currency amount',
+			notes: 'Choose to show the coin or token amount',
 		}
 	)
 	.add(
@@ -257,5 +286,26 @@ storiesOf('BadgerBadge', module)
 		{
 			notes:
 				'if watchAddress is true, the payment will turn to confirmed when the address receives a payment from any source.  Including other people.  This is ideal to use if the payment codes are unique for the checkout.  Not great if the payment address is shared by users.',
+		}
+	)
+	.add(
+		'controlled step',
+		() => (
+			<BadgerBadge
+				amount={0.0001}
+				to={text(
+					'To Address',
+					'bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g'
+				)}
+				stepControlled={select(
+					'step',
+					['fresh', 'pending', 'complete'],
+					'fresh'
+				)}
+			/>
+		),
+		{
+			notes:
+				'Controlled step overrides the component step state.  Valuable for payment systems where the app/backend does payment confirmation.',
 		}
 	);