+++ src/components/index.js
@@ -2,3 +2,4 @@
 
 export { default as BadgerButton } from './BadgerButton';
 export { default as BadgerBadge } from './BadgerBadge';
+export { default as PriceDisplay } from './PriceDisplay';