+++ src/hoc/BadgerBase/index.js
@@ -3,8 +3,9 @@
 import BadgerBase, {
 	type ButtonStates,
 	type BadgerBaseProps,
+	type ValidCoinTypes,
 } from './BadgerBase';
 
-export type { ButtonStates, BadgerBaseProps };
+export type { ButtonStates, BadgerBaseProps, ValidCoinTypes };
 
 export default BadgerBase;